migrate: SQLA 2.0

+++ flask_roots/cli/migrate.py
@@ -34,7 +34,7 @@ def main(dry_run, verbose, echo, no_backup):
             )
             if verbose:
                 print(f"Backing up SQLite:\n\tcp {dbpath}\n\tto {backup}")
-            if not dry_run:
+            if not dry_run and os.path.exists(dbpath):
                 shutil.copy(dbpath, backup)
 
 
@@ -49,12 +49,15 @@ def main(dry_run, verbose, echo, no_backup):
     table.create(checkfirst=True, bind=engine)
 
     root_path = flask.current_app.root_path
-    print(f"Root path: {root_path}")
 
     with engine.connect() as con:
         applied = dict(list(con.execute(sa.select(table.c.name, table.c.time))))
+    # For older naming which retained the extension.
+    applied = {k.replace('.py:', ':'): v for k, v in applied.items()}
 
     patchdir = os.path.join(root_path, 'etc', 'schema')
+    print(f"Schema directory: {patchdir}")
+
     for dirpath, dirnames, filenames in os.walk(patchdir, followlinks=True):
 
         # Need to sort this since Heroku lists them out of order.
@@ -62,7 +65,7 @@ def main(dry_run, verbose, echo, no_backup):
 
             if not filename.endswith('.py'):
                 continue
-
+            
             fullname = os.path.join(dirpath, filename)
             relname = os.path.relpath(fullname, patchdir)
             basename, ext = os.path.splitext(relname)
@@ -77,10 +80,13 @@ def main(dry_run, verbose, echo, no_backup):
             if upgrade and upgrade not in patches:
                 patches.append(upgrade)
             
-
+            if not patches:
+                cont
+            
             for patch in patches:
+
                 name = patch.__name__
-                patch_name = relname + ':' + name
+                patch_name = basename + ':' + name
                 
                 try:
                     apply_time = applied.pop(patch_name)
@@ -88,23 +94,23 @@ def main(dry_run, verbose, echo, no_backup):
                 except KeyError:
 
                     if dry_run:
-                        print(patch_name)
-                        print('\twould apply')
+                        print(patch_name, "would apply")
                     else:
-                        print(patch_name)
-                        print('\tapplying...',)
+                        print(patch_name, "applying...", )
 
                         patch(engine)
 
-                        engine.execute(table.insert(), name=patch_name)
+                        with engine.begin() as con:
+                            con.execute(table.insert(), dict(name=patch_name))
+                        
                         print('Done.')
 
                 else:
                     if verbose:
-                        print(patch_name)
-                        print('\tapplied', apply_time.isoformat(' '))
+                        print(patch_name, 'applied', apply_time.isoformat(' '))
     
-    print(f"{len(applied)} orphan patches")
+    if applied:
+        print(f"{len(applied)} orphaned patches:")
     for name, time in sorted(applied.items(), key=lambda x: (x[1], x[0])):
-        print(f"\t{time} {name}")
+        print(f"{name} applied {time.isoformat(' ')} (orphaned)")
